Fix warnings and errors for gcc11

+++ src/connection.cpp
@@ -252,7 +252,7 @@ void Connection::acceptConnection()
 {
 	try{
 		++m_pendingRead;
-		m_readTimer.expires_from_now(boost::posix_time::seconds(Connection::read_timeout));
+		m_readTimer.expires_from_now(boost::posix_time::seconds((long)Connection::read_timeout));
 		m_readTimer.async_wait( boost::bind(&Connection::handleReadTimeout, boost::weak_ptr<Connection>(shared_from_this()), boost::asio::placeholders::error));
 
 		// Read size of the first packet
@@ -304,7 +304,7 @@ void Connection::parseHeader(const boost::system::error_code& error)
 
 	try{
 		++m_pendingRead;
-		m_readTimer.expires_from_now(boost::posix_time::seconds(Connection::read_timeout));
+		m_readTimer.expires_from_now(boost::posix_time::seconds((long)Connection::read_timeout));
 		m_readTimer.async_wait( boost::bind(&Connection::handleReadTimeout, boost::weak_ptr<Connection>(shared_from_this()),
 			boost::asio::placeholders::error));
 
@@ -366,7 +366,7 @@ void Connection::parsePacket(const boost::system::error_code& error)
 
 	try{
 		++m_pendingRead;
-		m_readTimer.expires_from_now(boost::posix_time::seconds(Connection::read_timeout));
+		m_readTimer.expires_from_now(boost::posix_time::seconds((long)(long)Connection::read_timeout));
 		m_readTimer.async_wait( boost::bind(&Connection::handleReadTimeout, boost::weak_ptr<Connection>(shared_from_this()),
 			boost::asio::placeholders::error));
 
@@ -429,7 +429,7 @@ void Connection::internalSend(OutputMessage_ptr msg)
 
 	try{
 		++m_pendingWrite;
-		m_writeTimer.expires_from_now(boost::posix_time::seconds(Connection::write_timeout));
+		m_writeTimer.expires_from_now(boost::posix_time::seconds((long)Connection::write_timeout));
 		m_writeTimer.async_wait( boost::bind(&Connection::handleWriteTimeout, boost::weak_ptr<Connection>(shared_from_this()),
 			boost::asio::placeholders::error));
